Fixed test cases. Implemented kermit CRC16 (which I actually needed... I'm assuming that whoever wrote this library only got around to XModem because that's the one they needed.)
Edited the README for factual accuracy.  Expecting to find the library in a working state for 50+ algorithms is a bit of a gut-punch.

Incidentally, if you need one of the other table-based CRC16's implemented, it is rather easy using the tables from https://crccalc.com
Drop me a message and I'll try to bang it out for you.